Move version info plugin to convention plugin
