+++ kt/api-generator/settings.gradle.kts
@@ -1,17 +1,12 @@
 
 
 pluginManagement {
+    includeBuild("../build-logic")
     repositories {
         mavenLocal()
         mavenCentral()
         gradlePluginPortal()
         google()
-    }
-
-    resolutionStrategy.eachPlugin {
-        when (requested.id.id) {
-            "com.utopia-rise.godot-dependencies" -> useModule("com.utopia-rise:godot-convention-gradle-plugin:0.0.1")
-        }
     }
 }
 
@@ -27,12 +22,6 @@ dependencyResolutionManagement {
         create("libs") {
             from(files("../gradle/libs.versions.toml"))
         }
-    }
-}
-
-includeBuild("../plugins/godot-convention-gradle-plugin") {
-    dependencySubstitution {
-        substitute(module("com.utopia-rise:godot-convention-gradle-plugin")).using(project(":")) // assuming api-generator is the root project of api-generator/api-generator
     }
 }
