+++ kt/build-logic/convention/build.gradle.kts
@@ -1,4 +1,4 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+import org.apache.tools.ant.filters.ReplaceTokens
 
 plugins {
     alias(libs.plugins.kotlin.jvm)
@@ -21,3 +21,23 @@ gradlePlugin {
         }
     }
 }
+
+tasks {
+    @Suppress("UNUSED_VARIABLE")
+    val processResources by getting(Copy::class) {
+        outputs.upToDateWhen { false }
+        val tokens = mapOf(
+            "godot.kotlin.jvm.version" to libs.versions.godotKotlinJvm.get(),
+            "godot.version" to libs.versions.godot.get(),
+            "kotlin.version" to libs.versions.kotlin.get(),
+        )
+        from("src/main/resources") {
+            include("*.properties")
+            filter<ReplaceTokens>("tokens" to tokens)
+        }
+    }
+}
+
+// copy task needs an explicit duplication strategy set for gradle >= 7
+// see: https://docs.gradle.org/7.0.2/dsl/org.gradle.api.tasks.Copy.html#org.gradle.api.tasks.Copy:duplicatesStrategy
+tasks.withType<Copy> { duplicatesStrategy = DuplicatesStrategy.INCLUDE }
