+++ kt/build-logic/convention/src/main/resources/build.properties
