+++ kt/godot-library/build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
     alias(libs.plugins.kotlin.jvm)
     id("com.utopia-rise.api-generator")
     id("com.utopia-rise.godot-publish")
-    id("com.utopia-rise.godot-dependencies")
+    id("com.utopia-rise.versioninfo")
     alias(libs.plugins.shadowJar)
 }
