--- kt/plugins/godot-convention-gradle-plugin/settings.gradle.kts
@@ -1,24 +1,0 @@
-
-pluginManagement {
-    repositories {
-        mavenLocal()
-        mavenCentral()
-        gradlePluginPortal()
-        google()
-    }
-}
-
-dependencyResolutionManagement {
-    @Suppress("UnstableApiUsage")
-    repositories {
-        mavenLocal()
-        mavenCentral()
-        google()
-    }
-
-    versionCatalogs {
-        create("libs") {
-            from(files("../../gradle/libs.versions.toml"))
-        }
-    }
-}
