+++ kt/tools-common/settings.gradle.kts
@@ -1,16 +1,11 @@
 
 pluginManagement {
+    includeBuild("../build-logic")
     repositories {
         mavenLocal()
         mavenCentral()
         gradlePluginPortal()
         google()
-    }
-
-    resolutionStrategy.eachPlugin {
-        when (requested.id.id) {
-            "com.utopia-rise.godot-dependencies" -> useModule("com.utopia-rise:godot-convention-gradle-plugin:0.0.1")
-        }
     }
 }
 
@@ -26,11 +21,5 @@ dependencyResolutionManagement {
         create("libs") {
             from(files("../gradle/libs.versions.toml"))
         }
-    }
-}
-
-includeBuild("../plugins/godot-convention-gradle-plugin") {
-    dependencySubstitution {
-        substitute(module("com.utopia-rise:godot-convention-gradle-plugin")).using(project(":")) // assuming api-generator is the root project of api-generator/api-generator
     }
 }
